+++ structre/lib/src/reminds.dart
@@ -20,10 +20,8 @@ class DateRemindList extends ListBase<DateRemind> {
   DateRemindList([List<DateRemind> drl = const <DateRemind>[]]) : _drl = drl;
 
   /// Create list with [DateRemindObjectResolver]
-  factory DateRemindList.resolver(DateRemindObjectResolver resolver) {
-    List<DateRemind> dr = [];
-    resolver(dr);
-    return DateRemindList(dr);
+  DateRemindList._resolver(DateRemindObjectResolver resolver) : _drl = [] {
+    resolver(_drl);
   }
 
   @override
@@ -53,4 +51,37 @@ class DateRemindList extends ListBase<DateRemind> {
       .where((dr) => dr is Holiday)
       .map((e) => e as Holiday)
       .toList(growable: false);
+}
+
+class DateRemindListConversion {
+  /// Load a [List] of [source] from other classes
+  ///
+  /// Define [eventsCondition] and [holidayCondition] to check [S] is validate
+  /// structre to create [Events] or [Holiday].
+  /// Once it mathced, it will generated to [Events] or [Holiday] in
+  /// [exportEvents] or [exportHoliday]
+  static DateRemindList convertListFromSource<S>(List<S> source,
+          {required bool Function(S) eventsCondition,
+          required bool Function(S) holidayCondition,
+          required Events Function(S) exportEvents,
+          required Holiday Function(S) exportHoliday}) =>
+      DateRemindList._resolver((dr) {
+        source.forEach((srcdr) {
+          if (eventsCondition(srcdr))
+            dr.add(exportEvents(srcdr));
+          else if (holidayCondition(srcdr)) dr.add(exportHoliday(srcdr));
+        });
+      });
+
+  /// Getting [Events] and [Holiday] on different classes
+  ///
+  /// Unlike [convertListFromSource], it reduced the workflow of checking [Type]
+  /// and it will merge to [DateRemindList] after conversion completed.
+  static DateRemindList convertListFromDividedSource<E, H>(
+          List<E> eventsSource, List<H> holidaySource,
+          {required Events Function(E) exportEvents,
+          required Holiday Function(H) exportHoliday}) =>
+      DateRemindList._resolver((dr) => dr
+        ..addAll(eventsSource.map((e) => exportEvents(e)))
+        ..addAll(holidaySource.map((h) => exportHoliday(h))));
 }