+++ structre/lib/structre.dart
@@ -3,4 +3,10 @@ library cupertino_calendar_structre;
 
 export 'src/date.dart' hide getDTString, getDString;
 export 'src/week.dart';
-export 'src/reminds.dart' hide DateRemind, SingleDateRemind, DurationDateRemind;
+export 'src/reminds.dart'
+    hide
+        DateRemind,
+        SingleDateRemind,
+        DurationDateRemind,
+        DateRemindGenerator,
+        DateRemindListConversion;