+++ structre/lib/structre_tools.dart
@@ -1,0 +1,5 @@
+/// A tool that make more easier for generating object from third-party classes
+library cupertino_calendar_structre.tools;
+
+export 'src/reminds.dart'
+    show DateRemind, DateRemindGenerator, DateRemindListConversion;